# Improve space query performance