--- a/lib/benches/grounding_space.rs
+++ b/lib/benches/grounding_space.rs
@@ -10,21 +10,30 @@ use hyperon::space::grounding::*;
 fn space(size: isize) -> GroundingSpace {
     let mut space = GroundingSpace::new();
     for i in (0..size).step_by(1) {
-        space.add(expr!("A" {i}));
+        let func_sym = Atom::sym(format!("func-{:X}", i));
+        let func_def = Atom::expr([Atom::sym("="),
+            Atom::expr([func_sym, Atom::var("x")]),
+            Atom::var("x")]);
+        println!("func_def: {}", func_def);
+        space.add(func_def);
     }
     space
 }
 
 #[bench]
 fn query_x10(bencher: &mut Bencher) {
-
     let space = space(10);
-    bencher.iter(|| space.query(&expr!("A" {0})))
+    bencher.iter(|| {
+        let res = space.query(&expr!("=" ("func-9" "arg") X));
+        assert_eq!(res.len(), 1);
+    })
 }
 
 #[bench]
 fn query_x100(bencher: &mut Bencher) {
     let space = space(100);
-
-    bencher.iter(|| space.query(&expr!("A" {0})))
+    bencher.iter(|| {
+        let res = space.query(&expr!("=" ("func-2A" "arg") X));
+        assert_eq!(res.len(), 1);
+    })
 }