… python/tests/test_common.py
… python/tests/test_examples.py
… python/tests/test_grounding_space.py
… python/tests/test_minelogy.py
… python/tests/test_pln_tv.py
… python/tests/test_run_metta.py