Add index to make search of patterns inside of `GroundingSpace` faster.
Add benchmark to measure time difference between searching in space with 10 atoms and space with 100 atoms.
Benchmark results before index is implemented:
```
     Running benches/grounding_space.rs (target/release/deps/grounding_space-ce7d1e9199cc7107)

running 2 tests
test query_x10  ... bench:      13,899 ns/iter (+/- 156)
test query_x100 ... bench:     124,750 ns/iter (+/- 2,605)
```
Benchmark results after index is implemented:
```
     Running benches/grounding_space.rs (target/release/deps/grounding_space-ce7d1e9199cc7107)

running 2 tests
test query_x10  ... bench:       3,808 ns/iter (+/- 121)
test query_x100 ... bench:       3,755 ns/iter (+/- 72)
```
One significant thing is not implemented yet: when value is removed from index the key is not removed, so memory will leak when different atoms are added and removed frequently.